+++ google_cloud/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.2.1-wip
 
+- Add `projectIdFromCredentialsFile()` to automatically discover project ID from the 
+credentials JSON file.
 - Require `package:http` `^1.0.0`.
 - Require Dart 3.5